v1.0.0+dev base commit
First commit of PureTorch v1.0.0-beta. v1.0.0+dev is the base/foundational version of v1.0.0-beta. Most of v0.x-beta functionality is missing, but will soon be added before merging to the main branch. (Note the version 'v1.0.0+dev' is to stay the same until merged to the main branch. Any changes after that can increment the MINOR or PATCH versions)